chore: move constraintsParser to processOpenAPI directory
