+++ src/components/Parameter/OAParameter.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { getConstraints } from '../../lib/constraintsParser'
+import { getConstraints } from '../../lib/processOpenAPI/constraintsParser'
 import OACodeValue from '../Common/OACodeValue.vue'
 import OAParameterAttribute from './OAParameterAttribute.vue'
 import OAParameterExamples from './OAParameterExamples.vue'
